misc: Make build on macOS work
