+++ lang/funcs/core/sys/cpucount_fact_test.go
@@ -15,6 +15,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+// +build !darwin
+
 package coresys
 
 import (
